requirements.txt
streamlit
pillow
numpy
tensorflow==2.15.0
h5py